[win32] utils: move to platform folder
