+++ cmake/treedata/windowsstore/subdirs.txt
@@ -5,10 +5,10 @@ xbmc/platform/win10/peripherals        platform/win10/peripherals
 xbmc/platform/win10/powermanagement    platfrom/win10/powermanagement
 xbmc/platform/win10/storage            platfrom/win10/storage
 xbmc/platform/win32/filesystem         platform/win32/filesystem
+xbmc/platform/win32/utils              platform/win32/utils
 xbmc/input/touch                       input/touch
 xbmc/input/touch/generic               input/touch/generic
 xbmc/network/mdns                      network/mdns
-xbmc/utils/win32                       utils/win32
 xbmc/rendering/dx                      rendering/dx
 xbmc/threads/platform/win              threads/win
 xbmc/windowing/win10                   windowing/win10
