+++ xbmc/platform/win32/utils/CMakeLists.txt
@@ -4,4 +4,4 @@ set(HEADERS gpu_memcpy_sse4.h
             memcpy_sse2.h
             Win32InterfaceForCLog.h)
 
-core_add_library(utils_win32)
+core_add_library(platform_win32_utils)
