+++ xbmc/platform/win32/utils/Win32InterfaceForCLog.h
