+++ xbmc/utils/log.cpp
@@ -30,7 +30,7 @@
 #include "platform/posix/utils/PosixInterfaceForCLog.h"
 typedef class CPosixInterfaceForCLog PlatformInterfaceForCLog;
 #elif defined(TARGET_WINDOWS)
-#include "win32/Win32InterfaceForCLog.h"
+#include "platform/win32/utils/Win32InterfaceForCLog.h"
 typedef class CWin32InterfaceForCLog PlatformInterfaceForCLog;
 #endif
